fix(qa): split CLI JSON parsing from process runner
